Reorganize menu ui stuff
